reviewed: eclipse.platform.swt

+++ eclipse.platform.releng.aggregator/eclipse.platform.swt/bundles/org.eclipse.swt/Eclipse SWT/win32/org/eclipse/swt/widgets/Tree.java
@@ -64,7 +64,7 @@ import org.eclipse.swt.internal.win32.*;
  * <dt><b>Styles:</b></dt>
  * <dd>SINGLE, MULTI, CHECK, FULL_SELECTION, VIRTUAL, NO_SCROLL</dd>
  * <dt><b>Events:</b></dt>
- * <dd>Selection, DefaultSelection, Collapse, Expand, SetData, MeasureItem, EraseItem, PaintItem</dd>
+ * <dd>Selection, DefaultSelection, Collapse, Expand, SetData, MeasureItem, EraseItem, PaintItem, EmptinessChanged</dd>
  * </dl>
  * <p>
  * Note: Only one of the styles SINGLE and MULTI may be specified.
@@ -2046,7 +2046,10 @@ void createItem (TreeColumn column, int index) {
 	}
 }
 
-// For fast bulk insert, see comments for TreeItem#TreeItem(TreeItem,int,int)
+/**
+ * The fastest way to insert many items is documented in {@link TreeItem#TreeItem(Tree,int,int)}
+ * and {@link TreeItem#setItemCount}
+ */
 void createItem (TreeItem item, long hParent, long hInsertAfter, long hItem) {
 	int id = -1;
 	if (item != null) {
@@ -2066,9 +2069,8 @@ void createItem (TreeItem item, long hParent, long hInsertAfter, long hItem) {
 				shrink = true;
 				length = Math.max (4, items.length * 3 / 2);
 			}
-			TreeItem [] newItems = new TreeItem [length];
-			System.arraycopy (items, 0, newItems, 0, items.length);
-			items = newItems;
+
+			itemsGrowArray (length);
 		}
 		lastID = id + 1;
 	}
@@ -2112,13 +2114,14 @@ void createItem (TreeItem item, long hParent, long hInsertAfter, long hItem) {
 	if (hFirstItem == hFirstIndexOf && itemCount != -1) itemCount++;
 	if (hItem == 0) {
 		/*
-		* Bug in Windows.  When a child item is added to a parent item
-		* that has no children outside of WM_NOTIFY with control code
-		* TVN_ITEMEXPANDED, the tree widget does not redraw the +/-
-		* indicator.  The fix is to detect the case when the first
-		* child is added to a visible parent item and redraw the parent.
+		* Bug in Windows.  When a child item is added to a collapsed
+		* parent item that has no children, Tree does not draw [-]
+		* indicator and the parent item continues to look as if it
+		* has no children. Reportedly this doesn't happen when item
+		* is added during WM_NOTIFY(TVN_ITEMEXPANDED). The workaround
+		* is to force redraw parent item.
 		*/
-		if (fixParent) {
+		if (fixParent && (hParent != OS.TVI_ROOT)) {
 			if (getDrawing () && OS.IsWindowVisible (handle)) {
 				RECT rect = new RECT ();
 				if (OS.TreeView_GetItemRect (handle, hParent, rect, false)) {
@@ -2158,6 +2161,14 @@ void createItem (TreeItem item, long hParent, long hInsertAfter, long hItem) {
 		 Later, setRedraw(true) will update scrollbars once.
 		 */
 		if (getDrawing ()) updateScrollBar ();
+		/*
+		 If this is the first item added fire an EmptinessChanged event.
+		 */
+		if (item != null && id == 0) {
+			Event event = new Event ();
+			event.detail = 0;
+			sendEvent (SWT.EmptinessChanged, event);
+		}
 	}
 }
 
@@ -2641,6 +2652,15 @@ void destroyItem (TreeItem item, long hItem) {
 	 Later, setRedraw(true) will update scrollbars once.
 	 */
 	if (getDrawing ()) updateScrollBar ();
+
+	/*
+	 If this is the last item removed fire an EmptinessChanged event.
+	 */
+	if (count == 0) {
+		Event event = new Event ();
+		event.detail = 1;
+		sendEvent (SWT.EmptinessChanged, event);
+	}
 }
 
 @Override
@@ -3834,6 +3854,22 @@ boolean isUseWsBorder () {
 	return true;
 }
 
+int itemsGetFreeCapacity() {
+	int count = 0;
+	for (TreeItem item : items) {
+		if (item == null)
+			count++;
+	}
+
+	return count;
+}
+
+void itemsGrowArray (int newCapacity) {
+	TreeItem [] newItems = new TreeItem [newCapacity];
+	System.arraycopy (items, 0, newItems, 0, items.length);
+	items = newItems;
+}
+
 void redrawSelection () {
 	if ((style & SWT.SINGLE) != 0) {
 		long hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_CARET, 0);
@@ -4096,6 +4132,9 @@ public void setInsertMark (TreeItem item, boolean before) {
 
 /**
  * Sets the number of root-level items contained in the receiver.
+ * <p>
+ * The fastest way to insert many items is documented in {@link TreeItem#TreeItem(Tree,int,int)}
+ * and {@link TreeItem#setItemCount}
  *
  * @param count the number of items
  *
@@ -4109,25 +4148,49 @@ public void setInsertMark (TreeItem item, boolean before) {
 public void setItemCount (int count) {
 	checkWidget ();
 	count = Math.max (0, count);
-	long hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_ROOT, 0);
-	setItemCount (count, OS.TVGN_ROOT, hItem);
+	setItemCount (count, OS.TVI_ROOT);
 }
 
-void setItemCount (int count, long hParent, long hItem) {
+void setItemCount (int count, long hParent) {
+	// Investigate existing items and decide what to do
+	long itemInsertAfter = 0;
+	int  numInserted = 0;
+	long itemDeleteFrom = 0;
+	{
+		// Iterate to position #count and find prev/next items at this position
+		int itemCount = 0;
+		long itemPrev = OS.TVI_FIRST;
+		long itemNext = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_CHILD, hParent);
+		while (itemNext != 0 && itemCount < count)
+		{
+			itemPrev = itemNext;
+			itemNext = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, itemNext);
+			itemCount++;
+		}
+
+		if ((itemCount == count) && (itemNext == 0)) {
+			// Exactly 'count' items, no need to do anything.
+			return;
+		} else if (itemCount == count) {
+			// Too many items, going to delete some
+			itemDeleteFrom = itemNext;
+		} else if (itemNext == 0) {
+			// Counted all items, and there is not enough, going to insert some.
+			itemInsertAfter = itemPrev;
+			numInserted = count - itemCount;
+		}
+	}
+
 	boolean redraw = false;
 	if (OS.SendMessage (handle, OS.TVM_GETCOUNT, 0, 0) == 0) {
 		redraw = getDrawing () && OS.IsWindowVisible (handle);
 		if (redraw) OS.DefWindowProc (handle, OS.WM_SETREDRAW, 0, 0);
 	}
-	int itemCount = 0;
-	while (hItem != 0 && itemCount < count) {
-		hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, hItem);
-		itemCount++;
-	}
+
 	boolean expanded = false;
 	TVITEM tvItem = new TVITEM ();
 	tvItem.mask = OS.TVIF_HANDLE | OS.TVIF_PARAM;
-	if (!redraw && (style & SWT.VIRTUAL) != 0) {
+	if (!redraw && (style & SWT.VIRTUAL) != 0 && (hParent != OS.TVI_ROOT)) {
 		/*
 		* Bug in Windows.  Despite the fact that TVM_GETITEMSTATE claims
 		* to return only the bits specified by the stateMask, when called
@@ -4137,32 +4200,49 @@ void setItemCount (int count, long hParent, long hItem) {
 		int state = (int)OS.SendMessage (handle, OS.TVM_GETITEMSTATE, hParent, OS.TVIS_EXPANDED);
 		expanded = (state & OS.TVIS_EXPANDED) != 0;
 	}
-	while (hItem != 0) {
-		tvItem.hItem = hItem;
-		OS.SendMessage (handle, OS.TVM_GETITEM, 0, tvItem);
-		hItem = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, hItem);
-		TreeItem item = tvItem.lParam != -1 ? items [(int)tvItem.lParam] : null;
-		if (item != null && !item.isDisposed ()) {
-			item.dispose ();
-		} else {
-			releaseItem (tvItem.hItem, tvItem, false);
-			destroyItem (null, tvItem.hItem);
-		}
-	}
-	if ((style & SWT.VIRTUAL) != 0) {
-		for (int i=itemCount; i<count; i++) {
-			if (expanded) ignoreShrink = true;
-			createItem (null, hParent, OS.TVI_LAST, 0);
-			if (expanded) ignoreShrink = false;
+
+	if (itemDeleteFrom != 0) {
+		while (itemDeleteFrom != 0) {
+			tvItem.hItem = itemDeleteFrom;
+			OS.SendMessage (handle, OS.TVM_GETITEM, 0, tvItem);
+			itemDeleteFrom = OS.SendMessage (handle, OS.TVM_GETNEXTITEM, OS.TVGN_NEXT, itemDeleteFrom);
+			TreeItem item = tvItem.lParam != -1 ? items [(int)tvItem.lParam] : null;
+			if (item != null && !item.isDisposed ()) {
+				item.dispose ();
+			} else {
+				releaseItem (tvItem.hItem, tvItem, false);
+				destroyItem (null, tvItem.hItem);
+			}
 		}
 	} else {
-		shrink = true;
-		int extra = Math.max (4, (count + 3) / 4 * 4);
-		TreeItem [] newItems = new TreeItem [items.length + extra];
-		System.arraycopy (items, 0, newItems, 0, items.length);
-		items = newItems;
-		for (int i=itemCount; i<count; i++) {
-			new TreeItem (this, SWT.NONE, hParent, OS.TVI_LAST, 0);
+		// For performance reasons, reserve the necessary space in items[]
+		int freeCapacity = itemsGetFreeCapacity();
+		if (numInserted > freeCapacity)
+			itemsGrowArray (items.length + numInserted - freeCapacity);
+
+		// Note: on Windows, insert complexity is O(pos), so for performance
+		// reasons, all items are inserted at minimum possible position, that
+		// is, all at the same position.
+
+		if ((style & SWT.VIRTUAL) != 0) {
+			for (int i = 0; i < numInserted; i++) {
+				/*
+				 * Bug 206806: Windows sends 'TVN_GETDISPINFO' when item is
+				 * being inserted. This causes 'SWT.SetData' to be sent to
+				 * user code. If it happens to query 'getItemCount()', it will
+				 * get wrong number of items because we're still inserting. The
+				 * workaround is to temporarily suppress 'SWT.SetData'. Note
+				 * that the boolean flag is misleadingly used for multiple
+				 * purposes.
+				 */
+				if (expanded) ignoreShrink = true;
+				createItem (null, hParent, itemInsertAfter, 0);
+				if (expanded) ignoreShrink = false;
+			}
+		} else {
+			for (int i = 0; i < numInserted; i++) {
+				new TreeItem (this, SWT.NONE, hParent, itemInsertAfter, 0);
+			}
 		}
 	}
 	if (redraw) {
